# reduced allocations in getindex!